Time: 1 ms (100.00%) | Memory: 42.5 MB (92.93%) - LeetSync

+++ 441-arranging-coins/arranging-coins.java
@@ -7,7 +7,7 @@ class Solution {
         while (l < r) {
             long m = l + ( (r - l) >> 1);
 
-            long coins = (long) ( m * (m + 1) / 2);
+            long coins = ( m * (m + 1) / 2);
 
             if (coins <= n) l = m + 1;
             else r = m;
